fastapi-auth-custom/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pymysql==1.1.0
SQLAlchemy==2.0.20
pydantic[email]==2.5.0
pydantic==2.5.0
alembic==1.12.0
python-multipart==0.0.9